cleaner module structure
