+++ mqtt/mqtt-bridge/src/persist/persistor.rs
@@ -85,7 +85,7 @@ mod tests {
     use matches::assert_matches;
     use mqtt3::proto::{Publication, QoS};
 
-    use crate::persist::{persistor::Persistor, waking_state::WakingMap, Key};
+    use crate::persist::{persistor::Persistor, Key, WakingMap};
 
     #[tokio::test]
     async fn insert() {
